Fixed BGP AS number field size
In the doExecute(), one of the asn
number assignment was still happening
in the int format. Fixed it by changing
the data type to Long, keeping in line
with changes done in :
https://git.opendaylight.org/gerrit/#/c/48845/3

+++ vpnservice/bgpmanager/bgpmanager-impl/src/main/java/org/opendaylight/netvirt/bgpmanager/commands/Neighbor.java
@@ -86,7 +86,7 @@ public class Neighbor extends OsgiCommandSupport {
                     if (!Commands.isValid(asNum, Commands.Validators.INT, AS)) {
                         return null;
                     } else {
-                        asn = Integer.valueOf(asNum);
+                        asn = Long.valueOf(asNum);
                     }
                 }
                 bm.addNeighbor(nbrIp, asn);
